Solving #31 Remove references to WarFile Move and adjust tests for WarFile into GenericMojoTests Fix bug, missing "s" on  ".ebextensions" Delete WarFile and tests
