+++ src/main/java/com/jcabi/beanstalk/maven/plugin/AbstractBeanstalkMojo.java
@@ -178,8 +178,6 @@ abstract class AbstractBeanstalkMojo extends AbstractMojo {
         try {
             final ZipFile zipFile = this.createZipFile();
             this.validate(zipFile);
-            final WarFile warFile = this.createWarFile(zipFile);
-            this.validateWarFile(warFile);
             zipFile.close();
         } catch (final IOException ex) {
             throw new MojoFailureException(
@@ -214,31 +212,12 @@ abstract class AbstractBeanstalkMojo extends AbstractMojo {
     }
 
     /**
-     * Creates a {@link WarFile} out of the {@link ZipFile}.
-     * @param zipfile The zip file.
-     * @return WarFile the war file.
-     */
-    protected WarFile createWarFile(final ZipFile zipfile) {
-        return new WarFile(zipfile);
-    }
-
-    /**
      * Creates a {@link ZipFile} out of the war.
      * @return ZipFile the zip file
      * @throws IOException Thrown in case of IO error.
      */
     protected ZipFile createZipFile() throws IOException {
         return new ZipFile(this.war);
-    }
-
-    /**
-     * Validate {@link WarFile} given.
-     * @param warfile The war file
-     * @throws MojoFailureException Throw in case of validation error.
-     */
-    protected void validateWarFile(final WarFile warfile)
-        throws MojoFailureException {
-        warfile.checkEbextensionsValidity();
     }
 
     /**
@@ -335,23 +314,26 @@ abstract class AbstractBeanstalkMojo extends AbstractMojo {
     }
 
     /**
-     * Validate the war file.
-     *
+     * Verifies that the .ebextensions contains valid configuration file or
+     * files.
      * @param zip Zip file
-     * @throws org.apache.maven.plugin.MojoFailureException Thrown, if
-     *  validation fails.
+     * @throws org.apache.maven.plugin.MojoFailureException Thrown, if the
+     *  .ebextensions does not exist in the WAR file, is empty or one of its
+     *  files is neither valid JSON, nor valid YAML.
      */
     protected void validate(final ZipFile zip) throws MojoFailureException {
-        if (zip.getEntry(".ebextension") == null) {
+        if (zip.getEntry(".ebextensions") == null) {
             throw new MojoFailureException(
                 ".ebextensions directory does not exist in the WAR file"
             );
         }
         final Enumeration<? extends ZipEntry> entries = zip.entries();
+        int files = 0;
         while (entries.hasMoreElements()) {
             final ZipEntry entry = entries.nextElement();
             if (entry.getName().startsWith(".ebextensions/")
                 && !entry.isDirectory()) {
+                files += 1;
                 final String text = this.readFile(zip, entry);
                 if (this.validJson(text) || this.validYaml(text)) {
                     continue;
@@ -365,6 +347,11 @@ abstract class AbstractBeanstalkMojo extends AbstractMojo {
                     )
                 );
             }
+        }
+        if (files < 1) {
+            throw new MojoFailureException(
+                ".ebextensions contains no config files."
+            );
         }
     }
